html/semantics/forms/the-input-element/password.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>input type password</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#password-state-(type=password)">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<input type=password id=password value="password">
<script>
  var pass = document.getElementById('password');

  test(function(){
    assert_equals(pass.value, "password");
    pass.value = "  pass  word  ";
    assert_equals(pass.value, "  pass  word  ");
  }, "sanitization algorithm doesn't strip leading and trailing whitespaces");

  test(function(){
    pass.value = "pass\u000Aword";
    assert_equals(pass.value, "password");
    pass.value = "\u000Apassword\u000A";
    assert_equals(pass.value, "password");
    pass.value = "pass\u000Dword";
    assert_equals(pass.value, "password");
    pass.value = "\u000Dpassword\u000D";
    assert_equals(pass.value, "password");
  }, "sanitization algorithm strips line breaks");
</script>
